Increase number of reviews per page to 100
With such a high number of reviews per page it also makes sense to show
the count even when there is no need for paging.

This allows moderators to get a better overview of the content.

The number of items per page has been refactored into a class constant.
It should not change at runtime and this also makes it easier to 
reference in tests.