Mark campaign images for permanent storage
This requires us to properly mark uploaded files as permanent and 
register/deregister usage of them. Otherwise updates of existing
campaigns are not possible due to validation for managed_file form
elements.

To achieve this file fields have been added to the Campaign model and 
the Campaign repository have been updated to handle loading and saving
of campaign files. Corresponding logic has been removed from the
Campaign form.

File usage expects referenced types to be entities. That is not the
case for us but we still need to register usage. Consequently we remove
the file usage field + link from the files list view.